+++ src/main/java/com/enonic/lib/graphql/ExecutionResultMapper.java
@@ -9,7 +9,7 @@ import graphql.language.SourceLocation;
 import graphql.validation.ValidationError;
 import graphql.validation.ValidationErrorClassification;
 
-import com.enonic.lib.graphql.rx.Publisher;
+import com.enonic.lib.graphql.reactive.Publisher;
 import com.enonic.xp.script.serializer.MapGenerator;
 import com.enonic.xp.script.serializer.MapSerializable;
