+++ src/main/java/com/enonic/lib/graphql/reactive/Flowable.java
@@ -1,4 +1,4 @@
-package com.enonic.lib.graphql.rx;
+package com.enonic.lib.graphql.reactive;
 
 import java.util.function.Predicate;
