+++ src/main/java/com/enonic/lib/graphql/reactive/PublishProcessor.java
@@ -1,4 +1,4 @@
-package com.enonic.lib.graphql.rx;
+package com.enonic.lib.graphql.reactive;
 
 import java.util.concurrent.Flow;
 import java.util.concurrent.SubmissionPublisher;
