+++ src/main/java/com/enonic/lib/graphql/reactive/Publisher.java
@@ -1,4 +1,4 @@
-package com.enonic.lib.graphql.rx;
+package com.enonic.lib.graphql.reactive;
 
 import org.reactivestreams.Subscriber;
